internal/debug: move debug service to an internal package
This is in preparation for adding a /debug/dag handler.
